Refactor useDebounceFn to use TypeScript
This commit refactors the useDebounceFn hook in order to add TypeScript
type annotations. The current type of the 'fn' parameter is specified as
a function that takes any number of arguments and returns any value.
Similarly, the 'delay' parameter is specified as a number.

The refactor also replaces the usage of the 'useRef' hook. Rather than
accessing 'current' through destructuring, it is directly assigned to
'current'. Additionally, the 'timer' property of 'current' is defined as
either a NodeJS.Timeout or null.

These updates improve the reliability and maintainability of the
codebase, as type errors can be caught at compile-time. This makes it
easier to reason about the behavior of the hook and reduces the
likelihood of runtime errors.